social: fixed facebook social plugin and social js loading
The javascript loading does not have to be in the social.html file,
neither do the #fb-root div. This was causing issue #40. The #fb-root
div is used to create the facebook iframe, it can be anywhere in the
page.

The javascript code was moved to the base.html template, and if anyone
does not want to have the social plugins in a page, just override the
block js_social.

Last, but not least, I've fixed the buggy Facebook IIFE loading and
changed the g+ plugin to use a div instead of a brand new tag. One can
not use a new tag without declaring its namespace, and there wasn't any
namespace, and instead of declare one, I've just dropped the invalid tag
and used a div.

Fixes issue #40.