Isolate bootstrap generation from the after_install script.
This is the main change. On the previous version any change to the after_install script would require to regenerate the bootstrapV.M.py files. As @andrewsmedina pointed out, this raises the maintenance cost of the bootstrap strategy.

To make things simpler, I've moved all the venv-extension.py content to the gen.py script and redesigned it to delegate any after_install actions to the external script hooks.py.

By doing this, I've eliminated any change on the bootstrap*.py scripts. A bootstrap script would only be regenerated to upgrade the virtualenv version.