core: added help function for sending emails
This function uses a Mailer class that send asynchronous emails using
threads.

I would like to use this with user registration, but that would imply
changing the RagistrationForm implementation, which is not very hard,
but it is much work for now. Anyway, I would love to see someone doing
this.